[mempool] Isolate shared mempool proto
