+++ admission_control/admission-control-service/Cargo.toml
@@ -23,6 +23,7 @@ executable-helpers = { path = "../../common/executable-helpers" }
 grpc_helpers = { path = "../../common/grpc_helpers" }
 logger = { path = "../../common/logger" }
 mempool = { path = "../../mempool" }
+mempool-shared-proto = { path = "../../mempool/mempool-shared-proto" }
 metrics = { path = "../../common/metrics" }
 proto_conv = { path = "../../common/proto_conv" }
 storage_client = { path = "../../storage/storage_client" }
