+++ admission_control/admission_control_proto/build.rs
@@ -9,7 +9,7 @@
 fn main() {
     let proto_root = "src";
     let dependent_root = "../../types/src/proto";
-    let mempool_dependent_root = "../../mempool/src/proto/shared";
+    let mempool_dependent_root = "../../mempool/mempool-shared-proto/src/proto/";
 
     build_helpers::build_helpers::compile_proto(
         proto_root,
