+++ mempool/Cargo.toml
@@ -19,6 +19,7 @@ prost = "0.5.0"
 tokio = "0.1.22"
 ttl_cache = "0.4.2"
 
+mempool-shared-proto = { path = "mempool-shared-proto" }
 bounded-executor = { path = "../common/bounded-executor" }
 config = { path = "../config" }
 failure = { path = "../common/failure_ext", package = "failure_ext" }
