+++ mempool/src/core_mempool/unit_tests/common.rs
@@ -1,14 +1,12 @@
 // Copyright (c) The Libra Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::{
-    core_mempool::{CoreMempool, TimelineState, TxnPointer},
-    proto::shared::mempool_status::MempoolAddTransactionStatusCode,
-};
+use crate::core_mempool::{CoreMempool, TimelineState, TxnPointer};
 use config::config::NodeConfigHelpers;
 use crypto::ed25519::*;
 use failure::prelude::*;
 use lazy_static::lazy_static;
+use mempool_shared_proto::proto::mempool_status::MempoolAddTransactionStatusCode;
 use rand::{rngs::StdRng, SeedableRng};
 use std::{collections::HashSet, iter::FromIterator};
 use types::{
