+++ mempool/src/proto/mod.rs
@@ -4,10 +4,9 @@
 #![allow(bare_trait_objects)]
 #![allow(missing_docs)]
 
-use crate::proto::shared::*;
 use ::types::proto::*;
+use mempool_shared_proto::proto::mempool_status;
 
 pub mod mempool;
 pub mod mempool_client;
 pub mod mempool_grpc;
-pub mod shared;
