--- mempool/src/proto/shared/mod.rs
@@ -1,4 +1,0 @@
-// Copyright (c) The Libra Core Contributors
-// SPDX-License-Identifier: Apache-2.0
-
-pub mod mempool_status;
